refactor(base): 优化 `BaseReadViewModel` 书籍初始化与状态同步逻辑
- 重构 `upBook`、`loadBookInfo` 和 `loadChapterList`，统一通过 `inBookshelf` 状态管理书籍持久化。
- 修复 `ReadBookViewModel` 和 `ReadMangaViewModel` 在初始化及保存图片时的书籍对象引用问题。
- 优化 `BookInfoViewModel` 的数据加载流程，移除冗余的状态设置。
- 修正 `BookshelfViewModel` 导入书籍时 `type` 字段的类型转换。
- 在 `VideoViewModel` 中将初始化逻辑移至 `execute` 协程内执行。

+++ app/src/main/java/io/legado/app/ui/main/bookshelf/BookshelfViewModel.kt
@@ -123,9 +123,9 @@ class BookshelfViewModel(application: Application) : BaseViewModel(application) 
                                     "wordCount" -> book.wordCount = value
                                     "tocUrl" -> book.tocUrl = value
                                 }
-                            } else if (value is Int) {
+                            } else if (value is Long) {
                                 when (key) {
-                                    "type" -> book.type = value
+                                    "type" -> book.type = value.toInt()
                                 }
 
                             }
